Add exceptions to PatientService

+++ BLL/Services/PatientService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BLL.DTO;
+using BLL.Exceptions;
 using BLL.Interfaces;
 using DAL.Entities;
 using DAL.Interfaces;
@@ -32,18 +33,34 @@ namespace BLL.Services
         public async Task DeletePatient(int id)
         {
             var patient = await _unitOfWork.PatientRepository.GetByIdAsync(id);
+            if (patient == null)
+            {
+                throw new EntityNotFoundException(nameof(patient), id);
+            }
+
             _unitOfWork.PatientRepository.Delete(patient);
             await _unitOfWork.SaveAsync();
         }
 
         public async Task<IEnumerable<PatientDTO>> GetAllPatients()
         {
+            var patients = await _unitOfWork.PatientRepository.GetAllAsync();
+            if (patients == null)
+            {
+                throw new EntityNotFoundException(nameof(patients), 0);
+            }
+
             return _mapper.Map<IEnumerable<Patient>, IEnumerable<PatientDTO>>(await _unitOfWork.PatientRepository.GetAllAsync());
         }
 
         public async Task<PatientDTO> GetPatientById(int id)
         {
             var patient = await _unitOfWork.PatientRepository.GetByIdAsync(id);
+            if (patient == null)
+            {
+                throw new EntityNotFoundException(nameof(patient), id);
+            }
+
             var patientDTO = _mapper.Map<PatientDTO>(patient);
             return patientDTO;
         }
@@ -51,6 +68,10 @@ namespace BLL.Services
         public async Task<PatientDTO> UpdatePatient(int id, PatientDTO patientDTO)
         {
             var patient = await _unitOfWork.PatientRepository.GetByIdAsync(id);
+            if (patient == null)
+            {
+                throw new EntityNotFoundException(nameof(patient), id);
+            }
 
             patient.BirthDate = patientDTO.BirthDate;
             patient.FirstName = patientDTO.FirstName;
